Task 2  Problem Statement 1: MS Excel
Problem Statement 1: MS Excel

You will be provided with a dataset in MS Excel. You are supposed to analyze the data and create pivot tables and dashboards.

As a data analysis intern, you are provided with a general store transaction dataset. You are supposed to do the following operations.

Draw a Pivot table to get the different types of outlets and their counts
Draw a Pivot table to get the total sales for different outlets
Draw a Pivot table to find out the total sales of different item types in different outlet types
In the third Pivot table, filter out (remove) low-fat products 
Add slicers for item fat and item type in the last Pivot table

Dataset: General_store_dataset.xlsx